Added more tests
